requirements.txt
python-telegram-bot==12.8
nekos.py
pillow
requests
pymongo
dnspython